renamed common.ts to file_io.ts
