+++ source/rss/writer.ts
@@ -1,7 +1,7 @@
 import { rss_from_blog } from "./converter.ts";
 import { Blog } from "./types.ts";
 import { serialize } from "../deps.ts";
-import { write_file } from "../common.ts";
+import { write_file } from "../file_io.ts";
 
 
 export const write_rss = (blog: Blog, destination_path: string) => {
